+++ src/Acorn/Bootstrap/LoadConfiguration.php
@@ -2,10 +2,10 @@
 
 namespace Roots\Acorn\Bootstrap;
 
-use Symfony\Component\Finder\Finder;
-use Illuminate\Contracts\Foundation\Application;
-use Illuminate\Contracts\Config\Repository as RepositoryContract;
 use Illuminate\Config\Repository;
+use Illuminate\Contracts\Config\Repository as RepositoryContract;
+use Illuminate\Contracts\Foundation\Application;
+use Symfony\Component\Finder\Finder;
 
 
 class LoadConfiguration
@@ -42,10 +42,6 @@ class LoadConfiguration
         $app->detectEnvironment(function () use ($config) {
             return $config->get('app.env', defined('WP_ENV') ? \WP_ENV : 'production');
         });
-
-        date_default_timezone_set($config->get('app.timezone', 'UTC'));
-
-        mb_internal_encoding('UTF-8');
     }
 
     /**
